pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "lume-deploy"
dynamic = ["version"]
description = "A package for deploying AI models and more using lume inspired interfaces"
authors = [
    {"name" = "Mateusz Leputa"},
    {"name" = "Kathryn Baker"},
    {"name" = "Jesse Bellister"},
]
requires-python = "==3.11.*"
dependencies = [
    "argparse",
    "botorch",
    "tensorflow",
    "lume-model@git+https://github.com/slaclab/lume-model.git@2921e6583a6cfd49285833eb851b361aacf65b4c",
    "pandas>=2.0.1",
    "numpy<2.0",
    "keras==3.1.1",
    "mlflow==2.12.1",
    "boto3==1.34.65",
    "psycopg2-binary",
    "k2eg==0.2.7",
    "psutil==6.0.0",
    "colorlog==6.8.2",
    "p4p",
    "pyepics"
]

[tool.hatch.version]
source = "vcs"
path = "model_manager/src/_version.py"

[tool.hatch.build.hooks.vcs]
source = "vcs"
version-file = "model_manager/src/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["model_manager"]
only-packages = true

[tool.ruff.format]
preview = true
quote-style = "single"

[project.scripts]
model_manager = "model_manager.scripts.main:main"


